refactor(core): decouple passwordless verification flow w/o fixing tests
